domain update int to float

+++ domain/poProgressHeader.go
@@ -7,8 +7,8 @@ type PoProgressHeader struct {
 	Date     time.Time `json:"date" gorm:"column:date;default:NULL"`
 	Status   string    `json:"status" gorm:"column:status"`
 	IsEbapp  int       `json:"is_ebapp" gorm:"column:isebapp"`
-	LastProg int       `json:"last_prog" gorm:"last_prog"`
-	NewProg  int       `json:"new_prog" gorm:"new_prog"`
+	LastProg float32   `json:"last_prog" gorm:"last_prog"`
+	NewProg  float32   `json:"new_prog" gorm:"new_prog"`
 }
 
 type PoProgressHeaderResponse struct {
